feat: extract sort items to hook
